refactor: modularize code for improved organization and maintainability
This commit refactors the oasis.py script into multiple modules for better code organization, maintainability, and readability. The core functionality remains the same, but the code is now structured into separate modules for tools, Ollama management, embedding handling, analysis, and report generation. This modularization enhances code reusability and makes it easier to maintain and extend the project.
